proyecto-restaurante/requirements.txt
django==4.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
psycopg2==2.9.3
gunicorn==20.1.0
django-heroku==0.3.1
django-cors-headers==3.13.0